Move choices used types in a new service, add property type in EmailTemplateTest
