+++ src/Form/Type/EmailTemplateType.php
@@ -10,15 +10,13 @@ declare(strict_types=1);
 
 namespace BitBag\SyliusMailTemplatePlugin\Form\Type;
 
-use BitBag\SyliusMailTemplatePlugin\Entity\EmailTemplateInterface;
 use BitBag\SyliusMailTemplatePlugin\Form\Type\Translation\EmailTemplateTranslationType;
-use BitBag\SyliusMailTemplatePlugin\Provider\EmailCodesProviderInterface;
+use BitBag\SyliusMailTemplatePlugin\Service\EmailTemplateServiceInterface;
 use Sylius\Bundle\ResourceBundle\Form\Type\ResourceTranslationsType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Translation\DataCollectorTranslator;
 
 final class EmailTemplateType extends AbstractType
 {
@@ -40,16 +38,10 @@ final class EmailTemplateType extends AbstractType
 
     public const PREFIX_TRANS_TYPE = 'bitbag_sylius_mail_template_plugin.ui.type.';
 
-    private EmailCodesProviderInterface $emailCodesProvider;
+    private EmailTemplateServiceInterface $emailTemplateService;
 
-    private DataCollectorTranslator $dataCollectorTranslator;
-
-    public function __construct(
-        EmailCodesProviderInterface $emailCodesProvider,
-        DataCollectorTranslator $dataCollectorTranslator
-    ) {
-        $this->emailCodesProvider = $emailCodesProvider;
-        $this->dataCollectorTranslator = $dataCollectorTranslator;
+    public function __construct(EmailTemplateServiceInterface $emailTemplateService) {
+        $this->emailTemplateService = $emailTemplateService;
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options): void
@@ -57,7 +49,7 @@ final class EmailTemplateType extends AbstractType
         $builder
             ->add(self::TYPE_FIELD_NAME, ChoiceType::class, [
                 'label' => self::TEMPLATE_TYPE_LABEL,
-                'choices' => $this->getAvailableEmailTemplateTypes($options['data']),
+                'choices' => $this->emailTemplateService->getAvailableEmailTemplateTypes($options['data']),
                 'choice_translation_domain' => 'mail_template_type',
             ])
             ->add(self::STYLE_CSS_FIELD_NAME, TextareaType::class, [
@@ -78,18 +70,5 @@ final class EmailTemplateType extends AbstractType
     public function getBlockPrefix(): string
     {
         return self::BLOCK_PREFIX;
-    }
-
-    private function getAvailableEmailTemplateTypes(EmailTemplateInterface $emailTemplate): array
-    {
-        $types = $this->emailCodesProvider->provideWithLabelsNotUsedTypes();
-        if (null !== $emailTemplate->getId() && null !== $dataType = $emailTemplate->getType()) {
-            $replaceDataType = str_replace('_', ' ', $dataType);
-            $labelName = ucwords(strtolower($replaceDataType));
-
-            $types[$this->dataCollectorTranslator->trans($labelName, [], self::MAIL_TEMPLATE_TYPE_DOMAIN)] = $dataType;
-        }
-
-        return $types;
     }
 }
