+++ src/Resources/config/services/services.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <services>
+        <service id="bitbag_sylius_mail_template_plugin.service.email_template_service" class="BitBag\SyliusMailTemplatePlugin\Service\EmailTemplateService">
+            <argument type="service" id="bitbag_sylius_mail_template_plugin.provider.email_codes" />
+            <argument type="service" id="translator.data_collector" />
+        </service>
+    </services>
+</container>
